third_party/blink/web_tests/fast/xsl/exslt-node-set-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
Test that exslt:node-set() function is supported.

SUCCESS
